[NDK] Fix prototype of Nt/ZwQuerySystemInformation
svn path=/trunk/; revision=75606

+++ reactos/dll/win32/kernel32/client/time.c
@@ -534,7 +534,7 @@ GetSystemTimes(OUT LPFILETIME lpIdleTime OPTIONAL,
 {
     PSYSTEM_PROCESSOR_PERFORMANCE_INFORMATION ProcPerfInfo;
     LARGE_INTEGER TotalUserTime, TotalKernTime, TotalIdleTime;
-    SIZE_T BufferSize, ReturnLength;
+    ULONG BufferSize, ReturnLength;
     CCHAR i;
     NTSTATUS Status;
